Future changes commented, fixed GET method and CoAP to HTTP connection
Comments on code which is not understood.
GET method to obtain list of translators hub active was not working,
it was not possible to build a XML response from a Map, so now it
builds from String.
First test with the whole system working found bug when sending a
payload from CoAP to HTTP, connection settings have to be written
before connection establish, as getOuputstream connects.